+++ packages/core/src/demoDevices.ts
@@ -944,12 +944,12 @@ export async function createDemoDevices(matterbridge: Matterbridge): Promise<voi
   });
   await registerDevice(ep, 'EVSE', 'ENERGY-14-01');
 
-  ep = new Evse('EVSE Full Features', 'ENERGY-14-01-FF', {
-    id: 'EvseFullFeatures',
+  ep = new Evse('EVSE Complete', 'ENERGY-14-01-1', {
+    id: 'EvseComplete',
     number: EndpointNumber(14_01_1),
     tagList: [getSemtag(CommonNumberTag.Two)],
     stateOfCharge: 80,
-    batteryCapacity: 40_000_000,
+    batteryCapacity: 70_000_000,
     vehicleId: null,
     rfid: true,
     v2x: true,
@@ -957,7 +957,7 @@ export async function createDemoDevices(matterbridge: Matterbridge): Promise<voi
     absMinPower: -7_400_000,
     absMaxPower: 7_400_000,
   });
-  await registerDevice(ep, 'EVSE Full Features', 'ENERGY-14-01-FF');
+  await registerDevice(ep, 'EVSE Complete', 'ENERGY-14-01-1');
 
   ep = new WaterHeater('Water Heater', 'ENERGY-14-02', {
     id: 'WaterHeater',